Finish renaming file->files
