+++ src/atomate2/amset/files.py
@@ -8,7 +8,7 @@ from typing import Dict, Optional, Union
 
 from monty.serialization import dumpfn, loadfn
 
-from atomate2.common.file import copy_files, get_zfile, gunzip_files, rename_files
+from atomate2.common.files import copy_files, get_zfile, gunzip_files, rename_files
 from atomate2.settings import Settings
 from atomate2.utils.file_client import FileClient, auto_fileclient
 from atomate2.utils.path import strip_hostname
