+++ src/atomate2/common/files.py
